Look up stops by TransitLayer index, not position in pattern. #2884
I also changed the RaptorRouter so it attaches the Fare after
creating the Itinerary, not inside the ItineraryMapper.
