+++ src/test/java/org/opentripplanner/routing/fares/MultipleFareServiceTest.java
@@ -6,12 +6,12 @@ import java.util.Arrays;
 import com.conveyal.r5.otp2.api.path.Path;
 import junit.framework.TestCase;
 
+import org.opentripplanner.routing.algorithm.raptor.transit.TransitLayer;
 import org.opentripplanner.routing.algorithm.raptor.transit.TripSchedule;
 import org.opentripplanner.routing.core.Fare;
 import org.opentripplanner.routing.core.Fare.FareType;
 import org.opentripplanner.routing.core.WrappedCurrency;
 import org.opentripplanner.routing.services.FareService;
-import org.opentripplanner.routing.spt.GraphPath;
 
 /**
  * 
@@ -28,7 +28,7 @@ public class MultipleFareServiceTest extends TestCase {
         }
 
         @Override
-        public Fare getCost(Path<TripSchedule> path) {
+        public Fare getCost(Path<TripSchedule> path, TransitLayer transitLayer) {
             return fare;
         }
     }
@@ -53,41 +53,41 @@ public class MultipleFareServiceTest extends TestCase {
         FareService fs3 = new SimpleFareService(fare3);
 
         AddingMultipleFareService mfs = new AddingMultipleFareService(new ArrayList<FareService>());
-        Fare fare = mfs.getCost(null);
+        Fare fare = mfs.getCost(null, null);
         assertNull(fare);
 
         mfs = new AddingMultipleFareService(Arrays.asList(fs1));
-        fare = mfs.getCost(null);
+        fare = mfs.getCost(null, null);
         assertEquals(100, fare.getFare(FareType.regular).getCents());
         assertEquals(null, fare.getFare(FareType.student));
 
         mfs = new AddingMultipleFareService(Arrays.asList(fs2));
-        fare = mfs.getCost(null);
+        fare = mfs.getCost(null, null);
         assertEquals(140, fare.getFare(FareType.regular).getCents());
         assertEquals(120, fare.getFare(FareType.student).getCents());
 
         mfs = new AddingMultipleFareService(Arrays.asList(fs1, fs2));
-        fare = mfs.getCost(null);
+        fare = mfs.getCost(null, null);
         assertEquals(240, fare.getFare(FareType.regular).getCents());
         assertEquals(220, fare.getFare(FareType.student).getCents());
 
         mfs = new AddingMultipleFareService(Arrays.asList(fs2, fs1));
-        fare = mfs.getCost(null);
+        fare = mfs.getCost(null, null);
         assertEquals(240, fare.getFare(FareType.regular).getCents());
         assertEquals(220, fare.getFare(FareType.student).getCents());
 
         mfs = new AddingMultipleFareService(Arrays.asList(fs1, fs3));
-        fare = mfs.getCost(null);
+        fare = mfs.getCost(null, null);
         assertEquals(100, fare.getFare(FareType.regular).getCents());
         assertEquals(180, fare.getFare(FareType.student).getCents());
 
         mfs = new AddingMultipleFareService(Arrays.asList(fs3, fs1));
-        fare = mfs.getCost(null);
+        fare = mfs.getCost(null, null);
         assertEquals(100, fare.getFare(FareType.regular).getCents());
         assertEquals(180, fare.getFare(FareType.student).getCents());
 
         mfs = new AddingMultipleFareService(Arrays.asList(fs1, fs2, fs3));
-        fare = mfs.getCost(null);
+        fare = mfs.getCost(null, null);
         assertEquals(240, fare.getFare(FareType.regular).getCents());
         assertEquals(300, fare.getFare(FareType.student).getCents());
     }
